Rename system/processes to system/process.
