+++ x-pack/auditbeat/module/system/_meta/config.yml.tmpl
@@ -6,7 +6,7 @@
   metricsets:
     - host
     - packages
-    - processes
+    - process
     {{ if eq .GOOS "linux" -}}
     - socket
     - user
