+++ x-pack/auditbeat/module/system/process/process.go
@@ -2,7 +2,7 @@
 // or more contributor license agreements. Licensed under the Elastic License;
 // you may not use this file except in compliance with the Elastic License.
 
-package processes
+package process
 
 import (
 	"strconv"
@@ -23,7 +23,7 @@ import (
 
 const (
 	moduleName    = "system"
-	metricsetName = "processes"
+	metricsetName = "process"
 )
 
 func init() {
