+++ x-pack/auditbeat/module/system/process/process_test.go
@@ -2,7 +2,7 @@
 // or more contributor license agreements. Licensed under the Elastic License;
 // you may not use this file except in compliance with the Elastic License.
 
-package processes
+package process
 
 import (
 	"testing"
@@ -21,6 +21,6 @@ func TestData(t *testing.T) {
 func getConfig() map[string]interface{} {
 	return map[string]interface{}{
 		"module":     "system",
-		"metricsets": []string{"processes"},
+		"metricsets": []string{"process"},
 	}
 }
